refactor package diff computation into a library
